+++ src/DI/Container.php
@@ -89,19 +89,38 @@ class Container
         $className = self::cleanName($className);
 
         try {
-            $reflection = new \ReflectionClass($className);
-        } catch (\Exception $e) {
-            $message = "Could not create instance of '{$className}'";
+            $dependencies = self::getDependenciesFor($className, $createAll);
+        } catch (\Parable\DI\Exception $e) {
+            $message = $e->getMessage();
             if ($parentClassName) {
                 $message .= ", required by '{$parentClassName}'";
             }
+            throw new \Parable\DI\Exception($message);
+        }
+        return new $className(...$dependencies);
+    }
+
+    /**
+     * Retrieve and instantiate all dependencies for the provided $className
+     *
+     * @param string $className
+     *
+     * @return array
+     * @throws \Parable\DI\Exception
+     */
+    public static function getDependenciesFor($className, $createAll = false)
+    {
+        try {
+            $reflection = new \ReflectionClass($className);
+        } catch (\Exception $e) {
+            $message = "Could not create instance of '{$className}'";
             throw new \Parable\DI\Exception($message);
         }
 
         $construct = $reflection->getConstructor();
 
         if (!$construct) {
-            return new $className();
+            return [];
         }
 
         $parameters = $construct->getParameters();
@@ -124,7 +143,7 @@ class Container
                 $dependencies[] = self::get($subClassName, $className);
             }
         }
-        return new $className(...$dependencies);
+        return $dependencies;
     }
 
     /**